requirements.txt
# Core ML Libraries
xgboost>=2.0.0
lightgbm>=4.0.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0

# Data Processing
pyarrow>=12.0.0
fastparquet>=2023.4.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
tqdm>=4.65.0
joblib>=1.3.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Logging and Monitoring
mlflow>=2.5.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.25.0
